Extract test env classes where necessary
