+++ server/src/test/java/io/spine/server/model/given/DefaultStateRegistryTestEnv.java
@@ -18,26 +18,25 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.outbus.enrich;
+package io.spine.server.model.given;
 
-import com.google.common.testing.NullPointerTester;
-import com.google.protobuf.StringValue;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static io.spine.test.DisplayNames.NOT_ACCEPT_NULLS;
+import com.google.protobuf.Timestamp;
+import io.spine.server.entity.AbstractEntity;
 
 /**
  * @author Alexander Yevsyukov
+ * @author Dmytro Kuzmin
  */
-@DisplayName("SupportsFieldConversion should")
-class SupportsFieldConversionTest {
+public class DefaultStateRegistryTestEnv {
 
-    @Test
-    @DisplayName(NOT_ACCEPT_NULLS)
-    void passNullToleranceCheck() {
-        final SupportsFieldConversion predicate =
-                SupportsFieldConversion.of(StringValue.class, String.class);
-        new NullPointerTester().testAllPublicInstanceMethods(predicate);
+    /** Prevents instantiation of this utility class. */
+    private DefaultStateRegistryTestEnv() {
+    }
+
+    public static class TimerSnapshot extends AbstractEntity<Long, Timestamp> {
+
+        protected TimerSnapshot(Long id) {
+            super(id);
+        }
     }
 }
