+++ server/src/test/java/io/spine/server/model/given/MethodExceptionCheckerTestEnv.java
@@ -18,26 +18,39 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.outbus.enrich;
+package io.spine.server.model.given;
 
-import com.google.common.testing.NullPointerTester;
-import com.google.protobuf.StringValue;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import static io.spine.test.DisplayNames.NOT_ACCEPT_NULLS;
+import java.io.IOException;
 
 /**
- * @author Alexander Yevsyukov
+ * @author Dmytro Kuzmin
  */
-@DisplayName("SupportsFieldConversion should")
-class SupportsFieldConversionTest {
+public class MethodExceptionCheckerTestEnv {
 
-    @Test
-    @DisplayName(NOT_ACCEPT_NULLS)
-    void passNullToleranceCheck() {
-        final SupportsFieldConversion predicate =
-                SupportsFieldConversion.of(StringValue.class, String.class);
-        new NullPointerTester().testAllPublicInstanceMethods(predicate);
+    /** Prevents instantiation of this utility class. */
+    private MethodExceptionCheckerTestEnv() {
+    }
+
+    @SuppressWarnings("unused") // Reflective access.
+    public static class StubMethodContainer {
+
+        private static void methodNoExceptions() {
+        }
+
+        private static void methodCheckedException() throws Exception {
+            throw new IOException("Test checked exception");
+        }
+
+        private static void methodRuntimeException() throws RuntimeException {
+            throw new RuntimeException("Test runtime exception");
+        }
+
+        private static void methodCustomException() throws IOException {
+            throw new IOException("Test custom exception");
+        }
+
+        private static void methodDescendantException() throws IllegalStateException {
+            throw new IllegalStateException("Test descendant exception");
+        }
     }
 }
